+++ ECU2018/ECU2018/position.h
@@ -19,7 +19,7 @@ int encoderPosition_A();
 int encoderPosition_B();
 int encoderPosition_Z();
 int encoderPositionEngine();
-int encoderRPM();
+float encoderRPM();
 int encoderErrorCheck();
 
 #endif
